# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateDeptRequest:
    properties:
      deptTitle:
        description: 部门名称
        example: 技术部
        type: string
      pid:
        description: 父级部门ID，空表示顶级部门
        example: UUID
        type: string
      remark:
        description: 备注
        example: 技术部门
        type: string
      status:
        description: 状态 0=禁用 1=启用
        example: 1
        type: integer
    required:
    - deptTitle
    type: object
  models.CreateDictRequest:
    properties:
      color:
        description: 主题色
        example: '#2db7f5'
        type: string
      label:
        description: 字典标题
        example: 功能需求
        type: string
      pid:
        description: 父级字典ID
        example: UUID
        type: string
      remark:
        description: 备注
        example: 需求类型
        type: string
      status:
        description: 状态 0=禁用 1=启用
        example: 1
        type: integer
      type:
        description: 字典类型
        example: STORY_TYPE
        type: string
      value:
        description: 字典值
        example: "0"
        type: string
    required:
    - label
    - type
    type: object
...
  models.DeptTreeResponse:
    properties:
      children:
        description: 子部门
        items:
          $ref: '#/definitions/models.DeptTreeResponse'
        type: array
      createDate:
        description: 创建时间
        example: "2024-01-01 12:00:00"
        type: string
      deptId:
        description: 部门ID
        example: UUID
        type: string
      deptTitle:
        description: 部门名称
        example: 技术部
        type: string
      pid:
        description: 父级部门ID
        example: UUID
        type: string
      remark:
        description: 备注
        example: 技术部门
        type: string
      status:
        description: 状态 0=禁用 1=启用
        example: 1
        type: integer
    type: object
  models.DictTreeResponse:
    properties:
      children:
        description: 子字典
        items:
          $ref: '#/definitions/models.DictTreeResponse'
        type: array
      color:
        description: 主题色
        example: '#2db7f5'
        type: string
      createDate:
        description: 创建时间
        example: "2024-01-01 12:00:00"
        type: string
      id:
        description: 字典ID
        example: UUID
        type: string
      label:
        description: 字典标题
        example: 功能需求
        type: string
      pid:
        description: 父级字典ID
        example: UUID
        type: string
      remark:
        description: 备注
        example: 需求类型
        type: string
      status:
        description: 状态 0=禁用 1=启用
        example: 1
        type: integer
      type:
        description: 字典类型
        example: STORY_TYPE
        type: string
      updateDate:
        description: 更新时间
        example: "2024-01-01 12:00:00"
        type: string
      value:
        description: 字典值
        example: "0"
        type: string
    type: object
...
  models.Response:
    properties:
      code:
        description: 状态码 0=成功 -1=失败
        example: 0
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
      message:
        description: 提示信息
        example: ok
        type: string
    type: object
...
  models.UpdateDeptRequest:
    properties:
      deptTitle:
        description: 部门名称
        example: 技术部
        type: string
      pid:
        description: 父级部门ID
        example: UUID
        type: string
      remark:
        description: 备注
        example: 技术部门
        type: string
      status:
        description: 状态 0=禁用 1=启用
        example: 1
        type: integer
    required:
    - deptTitle
    type: object
  models.UpdateDictRequest:
    properties:
      color:
        description: 主题色
        example: '#2db7f5'
        type: string
      label:
        description: 字典标题
        example: 功能需求
        type: string
      pid:
        description: 父级字典ID
        example: UUID
        type: string
      remark:
        description: 备注
        example: 需求类型
        type: string
      status:
        description: 状态 0=禁用 1=启用
        example: 1
        type: integer
      type:
        description: 字典类型
        example: STORY_TYPE
        type: string
      value:
        description: 字典值
        example: "0"
        type: string
    required:
    - label
    - type
    type: object
...
host: localhost:9191
info:
  contact: {}
  description: Hive Admin Go 后端 API 接口文档
  title: Hive Admin API
  version: "1.0"
paths:
...
  /system/depts:
    delete:
      consumes:
      - application/json
      description: 批量删除部门
      parameters:
      - description: 部门ID列表
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除部门
      tags:
      - 系统管理-部门管理
    get:
      consumes:
      - application/json
      description: 获取部门树结构
      parameters:
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DeptTreeResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取部门树
      tags:
      - 系统管理-部门管理
    post:
      consumes:
      - application/json
      description: 创建新部门
      parameters:
      - description: 部门信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateDeptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建部门
      tags:
      - 系统管理-部门管理
  /system/depts/{deptId}:
    get:
      consumes:
      - application/json
      description: 根据部门ID获取部门详情
      parameters:
      - description: 部门ID
        in: path
        name: deptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DeptTreeResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取部门详情
      tags:
      - 系统管理-部门管理
    put:
      consumes:
      - application/json
      description: 更新部门信息
      parameters:
      - description: 部门ID
        in: path
        name: deptId
        required: true
        type: string
      - description: 部门信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDeptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新部门
      tags:
      - 系统管理-部门管理
  /system/depts/all:
    get:
      consumes:
      - application/json
      description: 获取所有部门（不分页）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DeptTreeResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有部门
      tags:
      - 系统管理-部门管理
  /system/dicts:
    delete:
      consumes:
      - application/json
      description: 批量删除字典
      parameters:
      - description: 字典ID列表
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除字典
      tags:
      - 系统管理-字典管理
    get:
      consumes:
      - application/json
      description: 获取字典树结构
      parameters:
      - description: 排序参数 排序时仅支持：label、type、value
        in: query
        name: sorts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DictTreeResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取字典树
      tags:
      - 系统管理-字典管理
    post:
      consumes:
      - application/json
      description: 创建新字典
      parameters:
      - description: 字典信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateDictRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建字典
      tags:
      - 系统管理-字典管理
  /system/dicts/{id}:
    get:
      consumes:
      - application/json
      description: 根据字典ID获取字典详情
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DictTreeResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取字典详情
      tags:
      - 系统管理-字典管理
    put:
      consumes:
      - application/json
      description: 更新字典信息
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: string
      - description: 字典信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDictRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新字典
      tags:
      - 系统管理-字典管理
  /system/dicts/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新字典启用/禁用状态
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新字典状态
      tags:
      - 系统管理-字典管理
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"